Backport 1c6fa113bd06585945e077f0d2004a291c27946d
